manim-cloud-renderer/requirements.txt
Flask==3.0.3
gunicorn==22.0.0
google-cloud-storage==2.17.0